Added docker Add comments to front
