--- client/my-app/src/services/AuthService.ts
@@ -1,20 +1,0 @@
-import $api from "../http";
-import {AxiosResponse} from "axios";
-import AuthResponse from "../models/responses/AuthResponse";
-
-class AuthService {
-    static async login(email: string, password: string)
-        : Promise<AxiosResponse<AuthResponse>> {
-        return $api.post<AuthResponse>("/login",{email, password});
-    }
-    static async registration(email: string, password: string)
-        : Promise<AxiosResponse<AuthResponse>> {
-        return $api.post<AuthResponse>("/registration",{email, password});
-    }
-
-    static async logout(): Promise<void> {
-        return $api.post("/logout");
-    }
-}
-
-export default AuthService;
