+++ client/src/models/IUser.ts
@@ -1,3 +1,7 @@
+/**
+ * @description - Модель пользователя
+ * @interface
+ */
 interface IUser {
     email: string,
     isActivated: boolean,
