+++ client/src/models/responses/AuthResponse.ts
@@ -1,5 +1,9 @@
 import IUser from "../IUser";
 
+/**
+ * @description - Модель ответа от сервера
+ * @interface
+ */
 interface AuthResponse {
     accessToken: string,
     refreshToken: string
