+++ server/.env
@@ -3,6 +3,7 @@ PORT=5000
 
 #Хост базы данных
 MONGO_HOST=127.0.0.1
+MONGO_HOST_DOCKER=host.docker.internal
 #MONGO_HOST=mongo
 
 #Порт к которому будет подключаться база данных
