+++ server/package.json
@@ -11,7 +11,7 @@
   "author": "l1msn",
   "license": "ISC",
   "dependencies": {
-    "bcrypt": "^5.0.1",
+    "bcryptjs": "^2.4.3",
     "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
     "dotenv": "^16.0.1",
@@ -21,6 +21,7 @@
     "mongodb": "^4.6.0",
     "mongoose": "^6.3.5",
     "nodemailer": "^6.7.5",
+    "npm": "^8.14.0",
     "uuid": "^8.3.2",
     "winston": "^3.7.2"
   },
