+++ server/services/userService.js
@@ -1,5 +1,5 @@
 //Инициализация библиотек
-const bcrypt = require("bcrypt");
+const bcrypt = require("bcryptjs");
 const uuid = require("uuid");
 require("dotenv").config();
